+++ examples/dump-pcr.rs
@@ -33,7 +33,7 @@ impl demultiplex::StreamConstructor for PcrDumpStreamConstructor {
             demultiplex::FilterRequest::ByPid(packet::Pid::PAT) => PcrDumpFilterSwitch::Pat(demultiplex::PatPacketFilter::default()),
             demultiplex::FilterRequest::Pmt{pid, program_number} => PcrDumpFilterSwitch::Pmt(demultiplex::PmtPacketFilter::new(pid, program_number)),
 
-            demultiplex::FilterRequest::ByStream(_, pmt_section, stream_info) => PcrDumpFilterSwitch::Pcr(PcrPacketFilter::construct(pmt_section, stream_info)),
+            demultiplex::FilterRequest::ByStream{ pmt, stream_info, .. } => PcrDumpFilterSwitch::Pcr(PcrPacketFilter::construct(pmt, stream_info)),
 
             demultiplex::FilterRequest::ByPid(_) => PcrDumpFilterSwitch::Null(demultiplex::NullPacketFilter::default()),
             demultiplex::FilterRequest::Nit{..} => PcrDumpFilterSwitch::Null(demultiplex::NullPacketFilter::default()),